Real Estate Web Page.
In this web site, the client will be able to look for properties, lots, rents, etc.
They can enter the house they're intrested and look for all the information of it. Learn from the company, ask them for help to sell or buy by completing a form in which the real state agent will recieve a notification on their email.